update hamburger menu, close icon still not appearing
